+++ environments/MyHeuristic.h
@@ -7,7 +7,7 @@
 template <class state>
 class MyHeuristic {
 public:
-	virtual int GetHCost(state &s) = 0;
+	virtual double GetHCost(state &s) = 0;
 };
 
 template <class state>
@@ -15,17 +15,17 @@ class MaxHeuristic
 {
 public:
 	MaxHeuristic(std::vector<MyHeuristic<state>*> h) :heurs(h) {}
-	int GetHCost(state &s);
+	double GetHCost(state &s);
 
 protected:
 	std::vector<MyHeuristic<state>*> heurs;
 };
 
 template<class state>
-int MaxHeuristic<state>::GetHCost(state & s)
+double MaxHeuristic<state>::GetHCost(state & s)
 {
-	int max = 0;
-	int hcost = 0;
+	double max = 0;
+	double hcost = 0;
 
 	for (int i = 0; i < heurs.size(); i++)
 	{
